+++ views/users/index.php
@@ -21,7 +21,7 @@
             <td><?=$user['Profile']?></td>
             <td>
                 <a href="<?=HOME_URI?>/users/edit/<?=$user['User_Id']?>" class="button small info mg-t-10">Editar</a>
-                <a href="<?=HOME_URI?>/users/delete/<?=$user['User_Id']?>" class="button small danger mg-t-10">Excluir</a>
+                <a href="<?=HOME_URI?>/users/delete/<?=$user['User_Id']?>" class="button small danger mg-t-10" onclick="return confirm('Tem certeza que deseja deletar este registro?')">Excluir</a>
             </td>
         </tr>
         <?php } ?>
